+++ portal-impl/src/com/liferay/portal/upgrade/v7_4_x/UpgradeListTypeCompanyId.java
@@ -5,7 +5,7 @@
 
 package com.liferay.portal.upgrade.v7_4_x;
 
-import com.liferay.portal.db.partition.DBPartitionUtil;
+import com.liferay.portal.db.partition.util.DBPartitionUtil;
 import com.liferay.portal.kernel.db.partition.DBPartition;
 import com.liferay.portal.kernel.model.ListType;
 import com.liferay.portal.kernel.security.auth.CompanyThreadLocal;
